added first version of channel support
