+++ test_requirements.txt
@@ -8,3 +8,4 @@ pytest-xdist>=1.14
 flake8>=3.0.4
 pylint>=1.5.4
 scipy
+jinja2
